+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-* Unit tests for HTML-rewriting logic.
+* Unit tests for HTML- and CSS-rewriting logic.
 ### Changed
 * Clean up the `WebResource` class's internal API.
 * Do not force a newline after the doctype in `HTMLRewriter.handle_decl()`.